// littleberry-ng/src/app/models/account.model.ts
export interface Account {
  SessionId: string;
  AccountId: number;
  FirstName: string;
  LastName: string;
  Phone: string;
  Email: string;
  AddressTitle: string;
  Address1: string;
  Address2: string;
  City: string;
  State: string;
  Zip: string;
  Country: string;
  FullAddress: string;
  AccountType: string;
  Status: string;
  CreateDate: string;
  ModifiedDate: string;
  ModifiedBy: string;
  CreateBy: string;
  Password: string;
  OfficeId: string;
}